# Optimize WireFormatConverter performance with 1.82x speedup