… core/src/main/scala/fastproto/WireFormatConverter.scala