## Summary
Optimize the interpreted WireFormatConverter to achieve a **1.82x speedup** (45% reduction in execution time) by eliminating boxing overhead and leveraging AbstractWireFormatConverter's optimized methods.

## Performance Results
**Before:**
- Direct WireFormatConverter: 4,524.811 ns/op
- Generated WireFormatConverter: 1,772.158 ns/op
- Performance gap: 2.55x

**After:**
- Direct WireFormatConverter: 2,492.905 ns/op (**1.82x speedup**)
- Generated WireFormatConverter: 1,822.695 ns/op (unchanged)
- Performance gap: 1.37x

## Key Optimizations

### 1. Type-Specific Primitive Lists
- Added `FloatList`, `DoubleList`, `BooleanList`, `ByteArrayList` Java classes
- Follows existing `IntList`/`LongList` pattern for optimal performance
- Direct primitive array access with count tracking

### 2. Eliminated Wrapper Classes
- Removed `FieldAccumulator` trait and all 6 wrapper implementations
- Direct list storage in `FieldMapping.accumulator: Any`
- One less pointer dereference per field access

### 3. Leveraged AbstractWireFormatConverter
- Extended `AbstractWireFormatConverter` instead of `BufferSharingRowConverter`
- Used optimized helper methods: `parsePackedInts/Longs/Floats/Doubles/Booleans`
- Imported static methods for clean code: `import fastproto.AbstractWireFormatConverter._`

### 4. Zero Boxing Overhead
- Type-specific accumulators for all primitive types
- Direct primitive array operations throughout
- Eliminated `ArrayBuffer[Any]` generic collections

## Technical Benefits
- **Memory efficiency**: Reduced allocations and GC pressure
- **Cache performance**: Better locality with primitive arrays vs object arrays  
- **JIT optimization**: Monomorphic calls instead of polymorphic dispatch
- **Code clarity**: Removed ~40 lines of wrapper boilerplate

## Test Coverage
✅ All 31 tests pass including:
- WireFormatConverter functionality tests
- Protobuf backport integration tests
- Complex message and repeated field handling

The interpreted WireFormatConverter now delivers performance within 37% of the generated version while maintaining full flexibility and compatibility.

🤖 Generated with [Claude Code](https://claude.ai/code)